flask_app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
reportlab
python-dotenv
email-validator
